完成后端代码开发

+++ dmall-service-impl/dmall-service-impl-search/src/main/java/com/dmall/search/service/impl/es/SkuEsDTO.java
@@ -60,12 +60,12 @@ public class SkuEsDTO implements Serializable {
     /**
      * sku评论数量
      */
-    private Long skuCommentCount;
+    private Integer skuCommentCount;
 
     /**
      * sku销量
      */
-    private Long skuSaleCount;
+    private Integer skuSaleCount;
 
     /**
      * sku上架时间
